requirements.txt
Pillow==8.3.2
requests==2.26.0